#4 fix: 패키지 구조 오류 정리
