+++ app/src/main/java/com/chunbae/narchive/presentation/ui/main/MainActivity.kt
@@ -1,17 +1,16 @@
-package com.chunbae.narchive.presentation.ui.presentation.ui.main
+package com.chunbae.narchive.presentation.ui.main
 
-import android.app.Activity
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import com.chunbae.narchive.R
 import com.chunbae.narchive.databinding.ActivityMainBinding
-import com.chunbae.narchive.presentation.ui.presentation.ui.main.calendar.CalendarFragment
-import com.chunbae.narchive.presentation.ui.presentation.ui.main.feed.FeedFragment
-import com.chunbae.narchive.presentation.ui.presentation.ui.main.group.GroupFragment
-import com.chunbae.narchive.presentation.ui.presentation.ui.main.settings.SettingsFragment
-import com.chunbae.narchive.presentation.ui.presentation.ui.main.todo.TodoFragment
+import com.chunbae.narchive.presentation.ui.main.calendar.CalendarFragment
+import com.chunbae.narchive.presentation.ui.main.feed.FeedFragment
+import com.chunbae.narchive.presentation.ui.main.group.GroupFragment
+import com.chunbae.narchive.presentation.ui.main.settings.SettingsFragment
+import com.chunbae.narchive.presentation.ui.main.todo.TodoFragment
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding : ActivityMainBinding
