+++ app/src/main/java/com/chunbae/narchive/presentation/ui/main/feed/FeedFragment.kt
@@ -1,4 +1,4 @@
-package com.chunbae.narchive.presentation.ui.presentation.ui.main.group
+package com.chunbae.narchive.presentation.ui.main.feed
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,17 +7,17 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import com.chunbae.narchive.R
-import com.chunbae.narchive.databinding.FragmentGroupBinding
+import com.chunbae.narchive.databinding.FragmentFeedBinding
 
-class GroupFragment : Fragment() {
-    private lateinit var binding : FragmentGroupBinding
+class FeedFragment : Fragment() {
+    private lateinit var binding : FragmentFeedBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_group, container, false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_feed, container, false)
         return binding.root
     }
 }
