+++ app/src/main/java/com/chunbae/narchive/presentation/ui/main/settings/SettingsFragment.kt
@@ -1,4 +1,4 @@
-package com.chunbae.narchive.presentation.ui.presentation.ui.main.todo
+package com.chunbae.narchive.presentation.ui.main.settings
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,17 +7,17 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import com.chunbae.narchive.R
-import com.chunbae.narchive.databinding.FragmentTodoBinding
+import com.chunbae.narchive.databinding.FragmentSettingsBinding
 
-class TodoFragment : Fragment() {
-    private lateinit var binding : FragmentTodoBinding
+class SettingsFragment : Fragment() {
+    private lateinit var binding : FragmentSettingsBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_todo, container, false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_settings, container, false)
         return binding.root
     }
 }
